update quantile huber, modify smoothing regressor and drop PDCD, delete unnecessary draft files
